Update Magic numbers.cpp

+++ Magic numbers.cpp
@@ -23,7 +23,7 @@ int main() {
 	long long int t,n,i;
 	cin>>t;
 	while(t--){
-	    int MOD=1e9+7;
+	    long long int MOD=1e9+7;
 	    cin>>n;
 	    for(i=0;i<n;i++)
 	        cin>>a[i];
